# docs/swagger-v1.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "SmartTaskManagement.API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/projects/{projectId}/tasks":  {
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "Tasks"
                                                                                ],
                                                                       "summary":  "Creates a new task inside the specified project.",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "projectId",
                                                                                              "in":  "path",
                                                                                              "description":  "Parent project identifier.",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "string",
                                                                                                             "format":  "uuid"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "requestBody":  {
                                                                                           "description":  "Task creation input.",
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/CreateTaskRequestDto"
                                                                                                                                               }
                                                                                                                                },
                                                                                                           "text/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/CreateTaskRequestDto"
                                                                                                                                        }
                                                                                                                         },
                                                                                                           "application/*+json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/CreateTaskRequestDto"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                       }
                                                                                       },
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK"
                                                                                                 }
                                                                                     }
                                                                   },
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "Tasks"
                                                                               ],
                                                                      "summary":  "Lists tasks for a specific project, with optional search, filtering, sorting, and pagination.",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "projectId",
                                                                                             "in":  "path",
                                                                                             "description":  "Parent project identifier.",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "uuid"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK"
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
                  "/api/tasks":  {
                                     "get":  {
                                                 "tags":  [
                                                              "Tasks"
                                                          ],
                                                 "summary":  "Lists tasks visible to the current user across all projects, with optional search, filtering, sorting, and pagination.",
                                                 "parameters":  [
                                                                    {
                                                                        "name":  "Search",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "type":  "string"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "Status",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "$ref":  "#/components/schemas/TaskItemStatus"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "Priority",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "$ref":  "#/components/schemas/TaskItemPriority"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "AssignedToUserId",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "type":  "string",
                                                                                       "format":  "uuid"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "DueDate",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "type":  "string",
                                                                                       "format":  "date-time"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "SortField",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "$ref":  "#/components/schemas/TaskSortField"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "SortDirection",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "$ref":  "#/components/schemas/SortDirection"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "PageNumber",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "format":  "int32"
                                                                                   }
                                                                    },
                                                                    {
                                                                        "name":  "PageSize",
                                                                        "in":  "query",
                                                                        "schema":  {
                                                                                       "type":  "integer",
                                                                                       "format":  "int32"
                                                                                   }
                                                                    }
                                                                ],
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK"
                                                                           }
                                                               }
                                             }
                                 },
                  "/api/tasks/{id}":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "Tasks"
                                                               ],
                                                      "summary":  "Retrieves a single task by id.",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "id",
                                                                             "in":  "path",
                                                                             "description":  "Task identifier.",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "string",
                                                                                            "format":  "uuid"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK"
                                                                                }
                                                                    }
                                                  },
                                          "put":  {
                                                      "tags":  [
                                                                   "Tasks"
                                                               ],
                                                      "summary":  "Updates an existing task\u0027s details.",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "id",
                                                                             "in":  "path",
                                                                             "description":  "Task identifier.",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "string",
                                                                                            "format":  "uuid"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "requestBody":  {
                                                                          "description":  "Update input.",
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/UpdateTaskRequestDto"
                                                                                                                              }
                                                                                                               },
                                                                                          "text/json":  {
                                                                                                            "schema":  {
                                                                                                                           "$ref":  "#/components/schemas/UpdateTaskRequestDto"
                                                                                                                       }
                                                                                                        },
                                                                                          "application/*+json":  {
                                                                                                                     "schema":  {
                                                                                                                                    "$ref":  "#/components/schemas/UpdateTaskRequestDto"
                                                                                                                                }
                                                                                                                 }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK"
                                                                                }
                                                                    }
                                                  },
                                          "delete":  {
                                                         "tags":  [
                                                                      "Tasks"
                                                                  ],
                                                         "summary":  "Soft-deletes a task. The task is hidden from normal queries.",
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "id",
                                                                                "in":  "path",
                                                                                "description":  "Task identifier.",
                                                                                "required":  true,
                                                                                "schema":  {
                                                                                               "type":  "string",
                                                                                               "format":  "uuid"
                                                                                           }
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     }
                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateTaskRequestDto":  {
                                                                    "type":  "object",
                                                                    "properties":  {
                                                                                       "title":  {
                                                                                                     "type":  "string",
                                                                                                     "nullable":  true
                                                                                                 },
                                                                                       "description":  {
                                                                                                           "type":  "string",
                                                                                                           "nullable":  true
                                                                                                       },
                                                                                       "status":  {
                                                                                                      "$ref":  "#/components/schemas/TaskItemStatus"
                                                                                                  },
                                                                                       "priority":  {
                                                                                                        "$ref":  "#/components/schemas/TaskItemPriority"
                                                                                                    },
                                                                                       "dueDate":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "date-time",
                                                                                                       "nullable":  true
                                                                                                   }
                                                                                   },
                                                                    "additionalProperties":  false
                                                                },
...
                                       "SortDirection":  {
                                                             "enum":  [
                                                                          0,
                                                                          1
                                                                      ],
                                                             "type":  "integer",
                                                             "format":  "int32"
                                                         },
...
                                       "TaskItemStatus":  {
                                                              "enum":  [
                                                                           0,
                                                                           1,
                                                                           2,
                                                                           3
                                                                       ],
                                                              "type":  "integer",
                                                              "format":  "int32"
                                                          },
...
                                   },
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter the JWT access token (without the \u0027Bearer \u0027 prefix).",
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ],
...
}